feat: dropzone, layout and member components
