+++ apps/web/src/components/Layout/Layout.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+export const layoutStyle = style({
+  display: 'flex',
+  flexDirection: 'column',
+  minHeight: '100vh',
+});
