--- apps/web/src/components/Layout/styles.ts
@@ -1,7 +1,0 @@
-import { styled } from 'stitches/stitches.config';
-
-export const StyledLayout = styled('div', {
-  minHeight: '100vh',
-  display: 'flex',
-  flexDirection: 'column',
-});
